+++ BankApplication/src/main/java/com/danny/BankApplication/service/alertTypesService.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.danny.BankApplication.exception.AlertTypeNotFoundException;
 import com.danny.BankApplication.model.alertTypes;
 import com.danny.BankApplication.repository.alertTypesRepo;
 
@@ -21,6 +22,10 @@ public class alertTypesService {
     }
 
     public Optional<alertTypes> getAlertType(int id){
+        Optional<alertTypes> alertType = repo.findById(id);
+        if (!alertType.isEmpty()) {
+            throw new AlertTypeNotFoundException("alert type is not found with id :"+id);
+        }
         return repo.findById(id);
     }
 
@@ -29,6 +34,9 @@ public class alertTypesService {
     }
 
     public void deleteAlertType(int id){
+        if(!repo.existsById(id)){
+            throw new AlertTypeNotFoundException("Alert type is not found with id :"+id);
+        }
         repo.deleteById(id);
     }
 }
