+++ BankApplication/src/main/java/com/danny/BankApplication/service/banksService.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.danny.BankApplication.exception.BankNotFoundException;
 import com.danny.BankApplication.model.banks;
 import com.danny.BankApplication.repository.bankRepo;
 
@@ -26,10 +27,17 @@ public class banksService {
     }
 
     public Optional<banks> getBank(int id){
+        Optional<banks> bank = repo.findById(id);
+        if (!bank.isEmpty()) {
+            throw new BankNotFoundException("Bank not found with id :"+id);
+        }
         return repo.findById(id);
     }
 
     public void deleteBank(int id){
+        if(!repo.existsById(id)){
+            throw new BankNotFoundException("Bank is not found with id :"+id);
+        }
         repo.deleteById(id);
     }
 }
