+++ BankApplication/src/main/java/com/danny/BankApplication/service/branchesService.java
@@ -3,10 +3,15 @@ package com.danny.BankApplication.service;
 import java.util.List;
 import java.util.Optional;
 
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.danny.BankApplication.exception.BankNotFoundException;
+import com.danny.BankApplication.exception.BranchNotFoundException;
+import com.danny.BankApplication.model.banks;
 import com.danny.BankApplication.model.branches;
+import com.danny.BankApplication.repository.bankRepo;
 import com.danny.BankApplication.repository.branchesRepo;
 
 
@@ -16,8 +21,17 @@ public class branchesService {
     @Autowired
     private branchesRepo repo;
 
+    @Autowired
+    private bankRepo bankRepo;
+
     public void saveBranch(branches branch){
-         repo.save(branch);
+        if(branch.getBank() != null && branch.getBank().getId() != null){
+            banks existingBank = bankRepo.findById(branch.getBank().getId())
+            .orElseThrow(() -> new BankNotFoundException("Bank not found with id :"+branch.getBank().getId()));
+
+            branch.setBank(existingBank);
+        }
+        repo.save(branch);
     }
 
     public List<branches> getAllBranches(){
@@ -25,11 +39,18 @@ public class branchesService {
     }
 
     public void deleteBranch(int id){
+        if(!repo.existsById(id)){
+            throw new BranchNotFoundException("Branch Not found with id :"+id);
+        }
         repo.deleteById(id);
     }
 
     public Optional<branches> getBranch(int id){
-        return repo.findById(id);
+        Optional<branches> branch = repo.findById(id);
+        if(!branch.isEmpty()){
+            throw new BranchNotFoundException("Branch Not found with id :"+id);
+        }
+        return branch;
     }
 
 }
