+++ BankApplication/src/main/java/com/danny/BankApplication/service/cardTypesService.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.danny.BankApplication.exception.CardTypeNotFoundException;
 import com.danny.BankApplication.model.cardTypes;
 import com.danny.BankApplication.repository.cardTypesRepo;
 
@@ -26,10 +27,17 @@ public class cardTypesService {
     }
 
     public void deleteCardTypes(int id){
+        if(!repo.existsById(id)){
+            throw new CardTypeNotFoundException("card type Not found with id :"+id);
+        }
         repo.deleteById(id);
     }
 
     public Optional<cardTypes> getCardTypes(int id){
+        Optional<cardTypes> cardType = repo.findById(id);
+        if (!cardType.isEmpty()) {
+            throw new CardTypeNotFoundException("card not found with id :"+id);
+        }
         return repo.findById(id);
     }
