+++ BankApplication/src/main/java/com/danny/BankApplication/service/loanOfficersService.java
@@ -6,7 +6,11 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.danny.BankApplication.exception.BranchNotFoundException;
+import com.danny.BankApplication.exception.LoanOfficerNotFoundException;
+import com.danny.BankApplication.model.branches;
 import com.danny.BankApplication.model.loanOfficers;
+import com.danny.BankApplication.repository.branchesRepo;
 import com.danny.BankApplication.repository.loanOfficersRepo;
 
 
@@ -16,15 +20,32 @@ public class loanOfficersService {
     @Autowired
     private loanOfficersRepo repo;
 
+    @Autowired
+    private branchesRepo branchRepo;
+
     public loanOfficers saveLoanOfficer(loanOfficers loanOfficer){
+        if(loanOfficer.getBranch() != null && loanOfficer.getBranch().getId() != null){
+            branches existingBranch = branchRepo.findById(loanOfficer.getBranch().getId())
+            .orElseThrow(() -> new BranchNotFoundException("Branch not found with id :"+loanOfficer.getBranch().getId()));
+            
+            loanOfficer.setBranch(existingBranch);
+
+        }
         return repo.save(loanOfficer);
     }
 
     public void deleteLoanOfficer(int id){
+        if(!repo.existsById(id)){
+            throw new LoanOfficerNotFoundException("loan officer Not found with id :"+id);
+        }
         repo.deleteById(id);
     }
 
     public Optional<loanOfficers> getLoanOfficer(int id){
+        Optional<loanOfficers> loanOfficer = repo.findById(id);
+        if(loanOfficer.isEmpty()){
+            throw new LoanOfficerNotFoundException("Loan officer not found with id :"+id);
+        }
         return repo.findById(id);
     }
