+++ BankApplication/src/main/java/com/danny/BankApplication/service/transactionTypesService.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.danny.BankApplication.exception.TransactionTypeNotFoundException;
 import com.danny.BankApplication.model.transactionTypes;
 import com.danny.BankApplication.repository.transactionTypesRepo;
 
@@ -21,6 +22,10 @@ public class transactionTypesService {
     }
 
     public Optional<transactionTypes> getTransactionType(int id){
+        Optional<transactionTypes> transactionType = repo.findById(id);
+        if (!transactionType.isEmpty()) {
+            throw new TransactionTypeNotFoundException("Transaction Type not found with id :"+id);
+        }
         return repo.findById(id);
     }
 
@@ -29,6 +34,9 @@ public class transactionTypesService {
     }
 
     public void deleteTransactionTypes(int id){
+        if(!repo.existsById(id)){
+            throw new TransactionTypeNotFoundException("Transaction type Not found with id :"+id);
+        }
         repo.deleteById(id);
     }
